At most locations, we operate in listen-only mode.   -> digi.active = false      digi.beacon = false
Sometimes, we are behind an rx rf-amplifier, or use the rx-antenna togehether with
an other rx.
Then it's very important inhibit unexpeted or TX, i.e. after config loss. Else it
would seriously damage our hardware.
If there are bugs, or if new features will come in (i.E. initial beacon on boot up),
I suppose a variable
  tx_enabled in the 'lora' section.
which defaults to false.

This variable assures that digi.active and digi.beacon is set to false, txpower is
set to 0dBm, and that the tx-routine is never entered.

=> The new configuration variable in is-cfg.json in the
   "lora": {
   }
   section is
     tx_enabled = false

I.e.:
        ..
        "lora": {
                "frequency_rx": 433775000,
                "gain_rx": 0,
                "frequency_tx": 433775000,
                "power": 20,
                "spreading_factor": 12,
                "signal_bandwidth": 125000,
                "coding_rate4": 5,
                "tx_enabled" = false
        },
        ..

Set tx_enabled to true if you are sure you safely can tx on 70cm lora.

Signed-off-by: Thomas Osterried <dl9sau@darc.de>